+++ trabajot1/src/main/java/com/ppooii/trabajot1/controller/PersonaController.java
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.ppooii.trabajot1.Services.Interfaces.PersonaServicelmpl;
+import com.ppooii.trabajot1.Services.PersonaServicelmpl;
 import com.ppooii.trabajot1.entities.Persona;
 
 @RestController
